+++ ChangeLog.md
@@ -5,6 +5,7 @@ Release 3.2.1:
   - feature: add raw value of ADS1115 to webui and mqtt, so you can calibrate your sensor by yourself
   - feature: change online state msg format to json with more details
   - feature: add optional AP bssid to force connect to a specific AP
+  - feature: filter ADS1115 moisture values with mor samples and EMA filter to get more stable values
 
 Release 3.2.0:
     +++++++ This version is not OTA compatible with older versions +++++++